requirements.txt
colorama==0.4.6
fabric==3.2.2
fastapi==0.115.11
httpx==0.28.1
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.0.1
sqlmodel==0.0.24
starlette==0.46.1
pyyaml~=6.0.2
apscheduler~=3.10.4
uvicorn~=0.34.0